requirements.txt
fastapi==0.115.12
fitz==0.0.1.dev2
langchain==0.3.24
python_docx==1.1.2
Requests==2.32.3
uvicorn==0.34.2
